Visioncave1.1/visioncave/requirements.txt
frappe
opencv-python-headless
numpy
Pillow
